chore(environment): move rebuild command to Environment namespace
